Snackbar service implemented

+++ src/app/opponents/opponents.component.ts
@@ -1,12 +1,11 @@
 import { SearchBarService } from './../search-bar/search-bar.service';
 import { OpponentsService, Opponent } from './opponents.service';
 import { Component, OnInit } from '@angular/core';
-import { FormControl } from '@angular/forms';
-import { Observable } from 'rxjs';
-import { startWith, map } from 'rxjs/operators';
-import { MatDialog, MatSnackBar } from '@angular/material';
+import { Observable, Subscriber, Subscription } from 'rxjs';
+import { MatDialog } from '@angular/material';
 import { OpponentComponent } from './opponent/opponent.component';
 import { NoteComponent } from './note/note.component';
+import { SnackbarService } from '../snackbar/snackbar-message.component';
 
 @Component({
   selector: 'app-opponents',
@@ -19,8 +18,8 @@ export class OpponentsComponent implements OnInit {
   opponentsFiltered: Observable<Opponent[]>;
   opponentSelected: Opponent;
 
-  constructor(public dialog: MatDialog, private opponentsService: OpponentsService, 
-              private searchBarService: SearchBarService, private _snackBar: MatSnackBar) { }
+  constructor(public dialog: MatDialog, private opponentsService: OpponentsService,
+    private searchBarService: SearchBarService, private snackbarService: SnackbarService) { }
 
   ngOnInit() {
     this.getOpponents(true);
@@ -31,7 +30,7 @@ export class OpponentsComponent implements OnInit {
   onAddOpponent() {
     const dialogRef = this.dialog.open(OpponentComponent, {
       width: '450px', panelClass: 'dialog-full-screen-mobile',
-      data: { opponentId: this.opponents.length }
+      data: { opponentId: this.createOpponentId() }
     });
 
     dialogRef.afterClosed().subscribe((opponent: Opponent) => {
@@ -73,21 +72,37 @@ export class OpponentsComponent implements OnInit {
   }
 
   onDeleteOpponent(opponent: Opponent) {
-    this.opponentsService.deleteOpponent(opponent).subscribe(opponent => {
-      this.getOpponents(true);
-    }, (error) => {
-      this._snackBar.open( 'Oops! Something went wrong', null , {
-        duration: 2000,
+    const confirmSnackBar$: Subscription = this.openConfirmSnackBar(`Are you sure you want to delete ${opponent.name}`, 'confirm', 'snackbar-main-confirm')
+      .subscribe((confirm: boolean) => {
+        if (confirm) {
+          this.opponentsService.deleteOpponent(opponent).subscribe(opponent => {
+            this.getOpponents(true);
+          }, (error) => {
+            this.openSnackBar('Oops! Something went wrong', 'error');
+          });
+        }
+      }, (error) => {
+        console.error(error);
+      }, () => {
+        confirmSnackBar$.unsubscribe();
       });
-    });
   }
 
   onDeleteNote(opponent: Opponent, noteIndex: number) {
-    opponent.notes = opponent.notes.filter((note: string, index: number) => {
-      return noteIndex !== index;
-    });
-    this.opponentSelected = opponent;
-    this.updateOpponent();
+    const confirmSnackBar$: Subscription = this.openConfirmSnackBar(`Are you sure you want to delete this note?`, 'confirm', 'snackbar-main-confirm')
+      .subscribe((confirm: boolean) => {
+        if (confirm) {
+          opponent.notes = opponent.notes.filter((note: string, index: number) => {
+            return noteIndex !== index;
+          });
+          this.opponentSelected = opponent;
+          this.updateOpponent();
+        }
+      }, (error) => {
+        console.error(error);
+      }, () => {
+        confirmSnackBar$.unsubscribe();
+      });
   }
   //******************************************************* */
   private getOpponents(resetControl = false) {
@@ -98,9 +113,7 @@ export class OpponentsComponent implements OnInit {
         this.searchBarService.resetControl();
       }
     }, (error) => {
-      this._snackBar.open( 'Oops! Something went wrong', null , {
-        duration: 2000,
-      });
+      this.openSnackBar('Oops! Something went wrong', 'error');
     });
   }
 
@@ -108,9 +121,7 @@ export class OpponentsComponent implements OnInit {
     this.opponentsService.addOpponent(this.opponentSelected).subscribe((opponent: Opponent) => {
       this.getOpponents(true);
     }, (error) => {
-      this._snackBar.open( 'Oops! Something went wrong', null , {
-        duration: 2000,
-      });
+      this.openSnackBar('Oops! Something went wrong', 'error');
     });
   }
 
@@ -118,10 +129,27 @@ export class OpponentsComponent implements OnInit {
     this.opponentsService.updateOpponent(this.opponentSelected).subscribe((opponent: Opponent) => {
       this.getOpponents();
     }, (error) => {
-      this._snackBar.open( 'Oops! Something went wrong', null , {
-        duration: 2000,
-      });
+      this.openSnackBar('Oops! Something went wrong', 'error');
     });
   }
 
+  private createOpponentId(): number {
+    let result: number;
+    if (this.opponents && this.opponents.length > 0) {
+      result = this.opponents[this.opponents.length - 1].id; // Get the last id
+      result++; // Create a new one
+    } else {
+      result = 0; // First Opponent Id
+    }
+    return result;
+  }
+
+  private openConfirmSnackBar(message, type, className): Observable<boolean> {
+    this.snackbarService.status.next({ message: message, type: type, class: className });
+    return this.snackbarService.onConfirmSnackbar();
+  }
+
+  private openSnackBar(message, type) {
+    this.snackbarService.status.next({ message: message, type: type, class: `snackbar-main-${type}`, duration: 4000 });
+  }
 }
